+++ packages/medusa-payment-stripe/src/services/stripe-provider.js
@@ -213,10 +213,12 @@ class StripeProviderService extends PaymentService {
   async refundPayment(paymentData, amount) {
     const { id } = paymentData
     try {
-      return this.stripe_.refunds.create({
+      await this.stripe_.refunds.create({
         amount: parseInt(amount * 100),
         payment_intent: id,
       })
+
+      return "refunded"
     } catch (error) {
       throw error
     }