+++ packages/medusa/src/api/routes/admin/shipping-options/create-shipping-option.js
@@ -21,6 +21,7 @@ export default async (req, res) => {
         })
       )
       .optional(),
+    is_return: Validator.boolean().default(false),
   })
 
   const { value, error } = schema.validate(req.body)