+++ packages/medusa-fulfillment-webshipper/CHANGELOG.md
@@ -3,6 +3,44 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.9](https://github.com/medusajs/medusa/compare/medusa-fulfillment-webshipper@1.0.8...medusa-fulfillment-webshipper@1.0.9) (2020-10-21)
+
+
+### Bug Fixes
+
+* **medusa-fulfillment-webshipper:** adds return customs_lines ([#130](https://github.com/medusajs/medusa/issues/130)) ([3b1f142](https://github.com/medusajs/medusa/commit/3b1f1422757f374efa8f3af99046753473d3f745))
+
+
+
+
+
+## [1.0.8](https://github.com/medusajs/medusa/compare/medusa-fulfillment-webshipper@1.0.7...medusa-fulfillment-webshipper@1.0.8) (2020-10-15)
+
+**Note:** Version bump only for package medusa-fulfillment-webshipper
+
+
+
+
+
+## [1.0.7](https://github.com/medusajs/medusa/compare/medusa-fulfillment-webshipper@1.0.6...medusa-fulfillment-webshipper@1.0.7) (2020-10-15)
+
+**Note:** Version bump only for package medusa-fulfillment-webshipper
+
+
+
+
+
+## [1.0.6](https://github.com/medusajs/medusa/compare/medusa-fulfillment-webshipper@1.0.5...medusa-fulfillment-webshipper@1.0.6) (2020-10-14)
+
+
+### Features
+
+* return shipping and flow ([#125](https://github.com/medusajs/medusa/issues/125)) ([c1e821d](https://github.com/medusajs/medusa/commit/c1e821d9d4d33756c7309e5cf110d7aa9b67297d))
+
+
+
+
+
 ## [1.0.5](https://github.com/medusajs/medusa/compare/medusa-fulfillment-webshipper@1.0.4...medusa-fulfillment-webshipper@1.0.5) (2020-10-06)
 
 **Note:** Version bump only for package medusa-fulfillment-webshipper